requirements.txt
django
requests
beautifulsoup4
lxml
google-genai
python-dotenv
